# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing Cloud Hypervisor VMs.",
        "title": "Cloud Hypervisor API",
        "termsOfService": "https://github.com/org/ch-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/org/ch-api/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms": {
            "get": {
                "description": "Returns a list of all virtual machines.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMs"
                ],
                "summary": "List VMs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.VMResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new virtual machine with the given configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMs"
                ],
                "summary": "Create VM",
                "parameters": [
                    {
                        "description": "VM configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateVMRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.VMResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}": {
            "get": {
                "description": "Returns the virtual machine with the given ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMs"
                ],
                "summary": "Get VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.VMResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the virtual machine with the given ID.",
                "tags": [
                    "VMs"
                ],
                "summary": "Delete VM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CPURequest": {
            "type": "object",
            "properties": {
                "boot_vcpus": {
                    "type": "integer"
                },
                "max_vcpus": {
                    "type": "integer"
                }
            }
        },
...
        "handler.ConsoleRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "handler.CreateVMRequest": {
            "type": "object",
            "properties": {
                "console": {
                    "$ref": "#/definitions/handler.ConsoleRequest"
                },
                "cpus": {
                    "$ref": "#/definitions/handler.CPURequest"
                },
                "disks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.DiskRequest"
                    }
                },
                "kernel": {
                    "$ref": "#/definitions/handler.KernelRequest"
                },
                "memory": {
                    "$ref": "#/definitions/handler.MemoryRequest"
                },
                "name": {
                    "type": "string"
                },
                "net": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.NetRequest"
                    }
                },
                "serial": {
                    "$ref": "#/definitions/handler.ConsoleRequest"
                }
            }
        },
        "handler.DiskRequest": {
            "type": "object",
            "properties": {
                "direct": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.KernelRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "handler.MemoryRequest": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "handler.NetRequest": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "tap": {
                    "type": "string"
                }
            }
        },
...
        "handler.VMResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/vmm.VmConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "errors": {
                    "description": "Errors holds machine-readable validation errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/problem.FieldError"
                    }
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code generated by the origin server.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        },
        "problem.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "vmm.BalloonConfig": {
            "type": "object",
            "properties": {
                "deflate_on_oom": {
                    "type": "boolean"
                },
                "free_page_reporting": {
                    "type": "boolean"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "vmm.CPUConfig": {
            "type": "object",
            "properties": {
                "boot_vcpus": {
                    "type": "integer"
                },
                "kvm_hyperv": {
                    "type": "boolean"
                },
                "max_vcpus": {
                    "type": "integer"
                },
                "topology": {
                    "$ref": "#/definitions/vmm.CPUTopology"
                }
            }
        },
        "vmm.CPUTopology": {
            "type": "object",
            "properties": {
                "cores_per_die": {
                    "type": "integer"
                },
                "dies_per_package": {
                    "type": "integer"
                },
                "packages": {
                    "type": "integer"
                },
                "threads_per_core": {
                    "type": "integer"
                }
            }
        },
        "vmm.ConsoleConfig": {
            "type": "object",
            "properties": {
                "mode": {
                    "description": "\"Off\", \"Tty\", \"File\", \"Socket\"",
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "vmm.DiskConfig": {
            "type": "object",
            "properties": {
                "direct": {
                    "type": "boolean"
                },
                "iommu": {
                    "type": "boolean"
                },
                "num_queues": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "pci_segment": {
                    "type": "integer"
                },
                "queue_size": {
                    "type": "integer"
                },
                "rate_limiter_group": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                },
                "vhost_socket": {
                    "type": "string"
                },
                "vhost_user": {
                    "type": "boolean"
                }
            }
        },
        "vmm.KernelConfig": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "vmm.MemoryConfig": {
            "type": "object",
            "properties": {
                "hotplug_method": {
                    "type": "string"
                },
                "hotplug_size": {
                    "type": "integer"
                },
                "hugepage_size": {
                    "type": "integer"
                },
                "hugepages": {
                    "type": "boolean"
                },
                "mergeable": {
                    "type": "boolean"
                },
                "shared": {
                    "type": "boolean"
                },
                "size": {
                    "type": "integer"
                },
                "zones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vmm.MemoryZoneConfig"
                    }
                }
            }
        },
        "vmm.MemoryZoneConfig": {
            "type": "object",
            "properties": {
                "host_numa_node": {
                    "type": "integer"
                },
                "hotplug_method": {
                    "type": "string"
                },
                "hotplug_size": {
                    "type": "integer"
                },
                "hugepage_size": {
                    "type": "integer"
                },
                "hugepages": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "mergeable": {
                    "type": "boolean"
                },
                "shared": {
                    "type": "boolean"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "vmm.NetConfig": {
            "type": "object",
            "properties": {
                "fd": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "host_mac": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "iommu": {
                    "type": "boolean"
                },
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "num_queues": {
                    "type": "integer"
                },
                "queue_size": {
                    "type": "integer"
                },
                "rate_limiter_group": {
                    "type": "string"
                },
                "tap": {
                    "type": "string"
                },
                "vhost_mode": {
                    "type": "string"
                },
                "vhost_socket": {
                    "type": "string"
                },
                "vhost_user": {
                    "type": "boolean"
                }
            }
        },
        "vmm.PayloadConfig": {
            "type": "object",
            "properties": {
                "cmdline": {
                    "type": "string"
                },
                "firmware": {
                    "type": "string"
                },
                "initramfs": {
                    "type": "string"
                },
                "kernel": {
                    "type": "string"
                }
            }
        },
        "vmm.RngConfig": {
            "type": "object",
            "properties": {
                "src": {
                    "type": "string"
                }
            }
        },
        "vmm.VmConfig": {
            "type": "object",
            "properties": {
                "balloon": {
                    "$ref": "#/definitions/vmm.BalloonConfig"
                },
                "console": {
                    "$ref": "#/definitions/vmm.ConsoleConfig"
                },
                "cpus": {
                    "$ref": "#/definitions/vmm.CPUConfig"
                },
                "disks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vmm.DiskConfig"
                    }
                },
                "kernel": {
                    "$ref": "#/definitions/vmm.KernelConfig"
                },
                "memory": {
                    "$ref": "#/definitions/vmm.MemoryConfig"
                },
                "net": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vmm.NetConfig"
                    }
                },
                "payload": {
                    "$ref": "#/definitions/vmm.PayloadConfig"
                },
                "rng": {
                    "$ref": "#/definitions/vmm.RngConfig"
                },
                "serial": {
                    "$ref": "#/definitions/vmm.ConsoleConfig"
                }
            }
        }
    },
...
}